pyproject.toml
[project]
name = "job"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apify-client>=2.5.0",
    "openai>=2.29.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
